requirements.txt
torch
torchvision
ptflops
thops
einops
torchsummaryX
EasyDict
opencv-python
shapely
Cython
scipy
pandas
pyyaml
json_tricks
scikit-image
yacs
tensorboardX
tqdm
ninja
onnx
onnxsim
